aircraft_api/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
celery==5.4.0
Django==5.1.1
django-celery-beat==2.7.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
pydantic==2.9.1
pydantic_core==2.23.3
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
alembic==1.13.1
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
distlib==0.3.7
Django==5.1.1
filelock==3.13.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
packaging==24.0
platformdirs==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.30
sqlparse==0.5.1
virtualenv==20.25.0
Werkzeug==3.0.3
python-decouple==3.8
uritemplate==4.1.1
